Relocate class `RouteWithUsage`
