+++ app/src/main/java/org/sedo/satmesh/model/rt/RouteEntry.java
@@ -2,7 +2,6 @@ package org.sedo.satmesh.model.rt;
 
 import androidx.annotation.NonNull;
 import androidx.room.ColumnInfo;
-import androidx.room.Embedded;
 import androidx.room.Entity;
 import androidx.room.Index;
 import androidx.room.PrimaryKey;
@@ -152,54 +151,5 @@ public class RouteEntry {
 				", hopCount=" + hopCount +
 				", lastUseTimestamp=" + lastUseTimestamp +
 				'}';
-	}
-
-	/**
-	 * Wrapper for {@link RouteEntry} and a date of usage (often the date of last usage).
-	 *
-	 * @author hsedo777
-	 */
-	public static class RouteWithUsage {
-		@Embedded
-		@NonNull
-		public RouteEntry routeEntry;
-		@Embedded(prefix = "usage_")
-		public RouteUsage routeUsage;
-		@Embedded(prefix = "backtracking_")
-		public RouteUsageBacktracking backtracking;
-
-		public RouteWithUsage(@NonNull RouteEntry routeEntry, RouteUsage routeUsage, RouteUsageBacktracking backtracking) {
-			this.routeEntry = routeEntry;
-			this.routeUsage = routeUsage;
-			this.backtracking = backtracking;
-		}
-
-		public boolean isWithoutUsage() {
-			return routeUsage == null && backtracking == null;
-		}
-
-		@Override
-		public boolean equals(Object o) {
-			if (o == null || getClass() != o.getClass()) return false;
-			RouteWithUsage that = (RouteWithUsage) o;
-			return Objects.equals(routeEntry, that.routeEntry) &&
-					Objects.equals(routeUsage, that.routeUsage) &&
-					Objects.equals(backtracking, that.backtracking);
-		}
-
-		@Override
-		public int hashCode() {
-			return Objects.hash(routeEntry, routeUsage, backtracking);
-		}
-
-		@NonNull
-		@Override
-		public String toString() {
-			return "RouteWithUsage{" +
-					"routeEntry=" + routeEntry +
-					", routeUsage=" + routeUsage +
-					", backtracking=" + backtracking +
-					'}';
-		}
 	}
 }
