+++ app/src/main/java/org/sedo/satmesh/nearby/NearbyRouteManager.java
@@ -13,11 +13,11 @@ import org.jetbrains.annotations.NotNull;
 import org.sedo.satmesh.model.Node;
 import org.sedo.satmesh.model.rt.BroadcastStatusEntry;
 import org.sedo.satmesh.model.rt.RouteEntry;
-import org.sedo.satmesh.model.rt.RouteEntry.RouteWithUsage;
 import org.sedo.satmesh.model.rt.RouteRequestEntry;
 import org.sedo.satmesh.model.rt.RouteUsage;
 import org.sedo.satmesh.model.rt.RouteUsageBacktracking;
 import org.sedo.satmesh.nearby.data.RouteRepository;
+import org.sedo.satmesh.nearby.data.RouteWithUsage;
 import org.sedo.satmesh.nearby.data.SelectionCallback;
 import org.sedo.satmesh.nearby.data.TransmissionCallback;
 import org.sedo.satmesh.proto.NearbyMessageBody;
